Make the test runner return a failure error code if a test failed.
For running in CI, I'd like to just know that without going through a
bunch of grepping.

+++ test/cts_runner-test-wrapper.sh
@@ -1,18 +1,26 @@
 #!/bin/sh
 
-set -e
-
 mode=$1
 shift
 runner=$1
 shift
 expected=$1
 shift
+expected_exitcode=$1
+shift
 
 output=`mktemp`
 
 $runner "$@" --output $output -- --deqp-mock-mode=$mode
+exitcode=$?
 
-diff -u $expected $output
+if ! diff -u $expected $output; then
+  exit 1
+fi
+
+if [ $exitcode != $expected_exitcode ]; then
+   echo "Runner returned $exitcode instead of $expected_exitcode"
+   exit 1
+fi
 
 rm $output
